+++ samplesheets/plugins.py
@@ -213,12 +213,12 @@ class ProjectAppPlugin(
         'igv_omit_bam': {
             'scope': SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT'],
             'type': 'STRING',
-            'label': 'BAM files to omit from IGV sessions',
+            'label': 'BAM and CRAM files to omit from IGV sessions',
             'default': '',
             'placeholder': ', '.join(settings.SHEETS_IGV_OMIT_BAM),
-            'description': 'Comma separated list of BAM file suffixes to omit '
-            'from generated IGV sessions. Overrides site-wide setting, affects '
-            'cancer and germline projects.',
+            'description': 'Comma separated list of BAM and CRAM file suffixes '
+            'to omit from generated IGV sessions. Overrides site-wide setting, '
+            'affects cancer and germline projects.',
             'user_modifiable': True,
         },
         'igv_omit_vcf': {