+++ samplesheets/studyapps/germline/tests/test_plugins_taskflow.py
@@ -3,6 +3,7 @@
 import os
 
 from django.conf import settings
+from django.test import override_settings
 from django.urls import reverse
 
 # Projectroles dependency
@@ -127,6 +128,29 @@ class TestGermlinePlugin(
             self.source_path, '{}_test.vcf.gz'.format(FAMILY_ID)
         )
         self.irods.data_objects.create(bam_path)
+        self.irods.data_objects.create(vcf_path)
+
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sc = self.plugin.get_shortcut_column(self.study, study_tables)
+
+        for i in range(0, 2):
+            self.assertEqual(sc['data'][i]['igv']['enabled'], True)
+            self.assertEqual(sc['data'][i]['files']['enabled'], True)
+        for i in range(3, 5):
+            self.assertEqual(sc['data'][i]['igv']['enabled'], False)
+            self.assertEqual(sc['data'][i]['files']['enabled'], False)
+
+    def test_get_shortcut_column_cram_file(self):
+        """Test get_shortcut_column() with cache item and CRAM file in iRODS"""
+        self.irods.collections.create(self.source_path)
+        cram_path = os.path.join(
+            self.source_path, '{}_test.cram'.format(SAMPLE_ID)
+        )
+        vcf_path = os.path.join(
+            self.source_path, '{}_test.vcf.gz'.format(FAMILY_ID)
+        )
+        self.irods.data_objects.create(cram_path)
         self.irods.data_objects.create(vcf_path)
 
         self.plugin.update_cache(self.cache_name, self.project)
@@ -266,14 +290,53 @@ class TestGermlinePlugin(
             + get_pedigree_file_path('vcf', self.source, study_tables),
         )
 
+    def test_get_shortcut_links_files_cram(self):
+        """Test get_shortcut_links() with CRAM file in iRODS"""
+        self.irods.collections.create(self.source_path)
+        cram_path = os.path.join(
+            self.source_path, '{}_test.cram'.format(SAMPLE_ID)
+        )
+        vcf_path = os.path.join(
+            self.source_path, '{}_test.vcf.gz'.format(FAMILY_ID)
+        )
+        self.irods.data_objects.create(cram_path)
+        self.irods.data_objects.create(vcf_path)
+
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sl = self.plugin.get_shortcut_links(
+            self.study, study_tables, family=[FAMILY_ID]
+        )
+
+        self.assertEqual(len(sl['data']['session']['files']), 1)
+        self.assertEqual(len(sl['data']['bam']['files']), 1)
+        self.assertEqual(len(sl['data']['vcf']['files']), 1)
+        self.assertEqual(
+            sl['data']['session']['files'][0]['url'],
+            reverse(
+                'samplesheets.studyapps.germline:igv',
+                kwargs={'genericmaterial': self.source.sodar_uuid},
+            ),
+        )
+        self.assertEqual(
+            sl['data']['bam']['files'][0]['url'],
+            settings.IRODS_WEBDAV_URL
+            + get_pedigree_file_path('bam', self.source, study_tables),
+        )
+        self.assertEqual(
+            sl['data']['vcf']['files'][0]['url'],
+            settings.IRODS_WEBDAV_URL
+            + get_pedigree_file_path('vcf', self.source, study_tables),
+        )
+
     def test_get_shortcut_links_multiple(self):
         """Test get_shortcut_links() with multiple BAM/VCF files"""
         self.irods.collections.create(self.source_path)
         bam_path = os.path.join(
             self.source_path, '{}_test_2022-11-06.bam'.format(SAMPLE_ID)
         )
-        bam_path2 = os.path.join(
-            self.source_path, '{}_test_2022-11-07.bam'.format(SAMPLE_ID)
+        cram_path = os.path.join(
+            self.source_path, '{}_test_2022-11-07.cram'.format(SAMPLE_ID)
         )
         vcf_path = os.path.join(
             self.source_path, '{}_test.vcf_2022-11-06.vcf.gz'.format(FAMILY_ID)
@@ -282,7 +345,7 @@ class TestGermlinePlugin(
             self.source_path, '{}_test.vcf_2022-11-07.vcf.gz'.format(FAMILY_ID)
         )
         self.irods.data_objects.create(bam_path)
-        self.irods.data_objects.create(bam_path2)
+        self.irods.data_objects.create(cram_path)
         self.irods.data_objects.create(vcf_path)
         self.irods.data_objects.create(vcf_path2)
 
@@ -293,17 +356,17 @@ class TestGermlinePlugin(
         )
 
         self.assertEqual(len(sl['data']['session']['files']), 1)
-        # Only the most recent bam and vcf should be returned
+        # Only the most recent bam/cram and vcf should be returned
         self.assertEqual(len(sl['data']['bam']['files']), 1)
         self.assertEqual(len(sl['data']['vcf']['files']), 1)
         self.assertTrue(
-            sl['data']['bam']['files'][0]['url'].endswith(bam_path2)
+            sl['data']['bam']['files'][0]['url'].endswith(cram_path)
         )
         self.assertTrue(
             sl['data']['vcf']['files'][0]['url'].endswith(vcf_path2)
         )
 
-    def test_get_shortcut_links_bam(self):
+    def test_get_shortcut_links_bam_only(self):
         """Test get_shortcut_links() with BAM file only"""
         self.irods.collections.create(self.source_path)
         bam_path = os.path.join(
@@ -319,7 +382,23 @@ class TestGermlinePlugin(
         self.assertEqual(len(sl['data']['bam']['files']), 1)
         self.assertEqual(len(sl['data']['vcf']['files']), 0)
 
-    def test_get_shortcut_links_vcf(self):
+    def test_get_shortcut_links_cram_only(self):
+        """Test get_shortcut_links() with CRAM file only"""
+        self.irods.collections.create(self.source_path)
+        cram_path = os.path.join(
+            self.source_path, '{}_test.cram'.format(SAMPLE_ID)
+        )
+        self.irods.data_objects.create(cram_path)
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sl = self.plugin.get_shortcut_links(
+            self.study, study_tables, family=[FAMILY_ID]
+        )
+        self.assertEqual(len(sl['data']['session']['files']), 1)
+        self.assertEqual(len(sl['data']['bam']['files']), 1)
+        self.assertEqual(len(sl['data']['vcf']['files']), 0)
+
+    def test_get_shortcut_links_vcf_only(self):
         """Test get_shortcut_links() with VCF file only"""
         self.irods.collections.create(self.source_path)
         vcf_path = os.path.join(
@@ -419,6 +498,28 @@ class TestGermlinePlugin(
         self.assertEqual(len(sl['data']['bam']['files']), 0)
         self.assertEqual(len(sl['data']['vcf']['files']), 0)
 
+    @override_settings(SHEETS_IGV_OMIT_BAM=['omit.cram'])
+    def test_get_shortcut_links_files_omit_cram(self):
+        """Test get_shortcut_links() with omittable CRAM file in iRODS"""
+        self.irods.collections.create(self.source_path)
+        cram_path_omit = os.path.join(
+            self.source_path, '{}_omit.cram'.format(SAMPLE_ID)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(FAMILY_ID)
+        )
+        self.irods.data_objects.create(cram_path_omit)
+        self.irods.data_objects.create(vcf_path_omit)
+
+        self.plugin.update_cache(self.cache_name, self.project)
+        study_tables = self.tb.build_study_tables(self.study)
+        sl = self.plugin.get_shortcut_links(
+            self.study, study_tables, family=[FAMILY_ID]
+        )
+        self.assertEqual(len(sl['data']['session']['files']), 0)
+        self.assertEqual(len(sl['data']['bam']['files']), 0)
+        self.assertEqual(len(sl['data']['vcf']['files']), 0)
+
     def test_get_shortcut_links_files_omit_override(self):
         """Test get_shortcut_links() with project-specific omit override"""
         app_settings.set(
@@ -483,6 +584,25 @@ class TestGermlinePlugin(
         self.assertEqual(ci['vcf'][FAMILY_ID], vcf_path)
         self.assertIsNone(ci['vcf'][FAMILY_ID2])
 
+    def test_update_cache_files_cram(self):
+        """Test update_cache() with CRAM file in iRODS"""
+        self.irods.collections.create(self.source_path)
+        cram_path = os.path.join(
+            self.source_path, '{}_test.cram'.format(SAMPLE_ID)
+        )
+        vcf_path = os.path.join(
+            self.source_path, '{}_test.vcf.gz'.format(FAMILY_ID)
+        )
+        self.irods.data_objects.create(cram_path)
+        self.irods.data_objects.create(vcf_path)
+        self.plugin.update_cache(self.cache_name, self.project)
+        ci = self.cache_backend.get_cache_item(
+            APP_NAME, self.cache_name, self.project
+        ).data
+        self.assertEqual(ci['bam'][self.source.name], cram_path)
+        self.assertEqual(ci['vcf'][FAMILY_ID], vcf_path)
+        self.assertIsNone(ci['vcf'][FAMILY_ID2])
+
     def test_update_cache_files_omit(self):
         """Test update_cache() with omittable files in iRODS"""
         self.irods.collections.create(self.source_path)
@@ -527,6 +647,25 @@ class TestGermlinePlugin(
             APP_NAME, self.cache_name, self.project
         ).data
         # Omitted files should not be returned
+        self.assertIsNone(ci['bam'][self.source.name])
+        self.assertIsNone(ci['vcf'][FAMILY_ID])
+
+    @override_settings(SHEETS_IGV_OMIT_BAM=['omit.cram'])
+    def test_update_cache_files_omit_only_cram(self):
+        """Test update_cache() with omittable CRAM file in iRODS"""
+        self.irods.collections.create(self.source_path)
+        cram_path_omit = os.path.join(
+            self.source_path, '{}_omit.cram'.format(SAMPLE_ID)
+        )
+        vcf_path_omit = os.path.join(
+            self.source_path, '{}_cnv.vcf.gz'.format(FAMILY_ID)
+        )
+        self.irods.data_objects.create(cram_path_omit)
+        self.irods.data_objects.create(vcf_path_omit)
+        self.plugin.update_cache(self.cache_name, self.project)
+        ci = self.cache_backend.get_cache_item(
+            APP_NAME, self.cache_name, self.project
+        ).data
         self.assertIsNone(ci['bam'][self.source.name])
         self.assertIsNone(ci['vcf'][FAMILY_ID])